requirements.txt
altgraph==0.17
amqp==2.5.2
apns2==0.7.1
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
coinbase==2.1.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decouple==0.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-celery==3.1.17
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-filter==2.2.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-push-notifications==2.0.0
Django-Pushy==1.0.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
download-gitignore==2.0.1
drf-writable-nested==0.6.0
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
h2==2.6.2
heroku==0.1.4
hpack==3.0.0
http-ece==1.1.0
httpimport==0.7.2
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
kombu==4.6.8
lazy-object-proxy==1.4.3
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1
MouseInfo==0.1.3
mysqlclient==1.4.6
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pefile==2019.4.18
phonenumbers==8.12.2
Pillow==7.1.1
psycopg2==2.8.5
pushjack==1.3.0
py-vapid==1.7.0
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.10.1
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pylint==2.5.2
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==1.5
python-decouple==3.3
python-dotenv==0.13.0
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2019.3
pywebpush==1.11.0
pywhatkit==3.8
pywin32==300
PyYAML==5.3.1
pyzbar==0.1.8
rave-python==1.1.2
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.1
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
tqdm==4.45.0
twilio==6.50.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
whois==0.9.8
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.12.1
xlwt==1.3.0
